+++ src/Page/GroupPage/GroupPageBarrel.ts
@@ -9,9 +9,11 @@ import LikeBind from '../../Type/LikeBind';
 import Comment from './Comment';
 import DivideBaS from '../../Component/Mission/DivideBa';
 import GroupActive from './GroupActive';
+// import Mind from './Mind';
 import GroupArticle from '../../Component/Mission/GroupArticle';
 
 export {
+  // Mind
   useState,
   useEffect,
   styled,
